feat: delete product
